india_bot/requirements.txt
ib_insync
yfinance
pandas
numpy
python-dotenv
pytz
requests
